pyproject.toml
[tool.poetry]
name = "polehammer-poster"
version = "1.0.0"
description = "A bot that knows quite a bit about chivalry 2"
authors = ["Jacob Barber <Jacoby6000@gmail.com>"]
license = "GPL3"
readme = "README.md"
packages = [{include = "polehammerposter", from="src/"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.7.2"
praw = "^7.7.1"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
types-requests = "^2.31.0.20240106"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
autoflake = "^2.2.0"
mypy = "^1.4.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [ "src" ]
